+++ srcs/ft_print_func/ft_print_func.h
@@ -6,7 +6,7 @@
 /*   By: rfriscca <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/05/30 15:54:00 by rfriscca          #+#    #+#             */
-/*   Updated: 2016/06/16 15:28:39 by rfriscca         ###   ########.fr       */
+/*   Updated: 2016/06/22 14:05:29 by rfriscca         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,12 +17,12 @@
 void	ft_putwchar(wchar_t c);
 void	ft_printwchar(wchar_t c, t_stock stock);
 void	ft_printchar(char c, t_stock stock);
-void	ft_printhexcaps(int nbr, t_stock stock);
-void	ft_printoctal(int nbr, t_stock stock);
+void	ft_printhexcaps(uintmax_t nbr, t_stock stock);
+void	ft_printoctal(uintmax_t nbr, t_stock stock);
 void	ft_printwstr(wchar_t *str, t_stock stock);
-void	ft_printhex(int nbr, t_stock stock);
+void	ft_printhex(uintmax_t nbr, t_stock stock);
 void	ft_printstr(char *str, t_stock stock);
-void	ft_printnbr(int nbr, t_stock stock);
-void	ft_printunbr(unsigned int nbr, t_stock stock);
+void	ft_printnbr(intmax_t nbr, t_stock stock);
+void	ft_printunbr(uintmax_t nbr, t_stock stock);
 
 #endif
