+++ srcs/ft_print_func/ft_printoctal.c
@@ -6,13 +6,13 @@
 /*   By: rfriscca <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/05/30 15:47:20 by rfriscca          #+#    #+#             */
-/*   Updated: 2016/06/07 13:03:13 by rfriscca         ###   ########.fr       */
+/*   Updated: 2016/06/22 13:53:28 by rfriscca         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft_printf.h"
 
-void	ft_printoctal(int nbr, t_stock stock)
+void	ft_printoctal(uintmax_t nbr, t_stock stock)
 {
 	int		spacenum;
 	int		size;
